refactor(scripts): ブラウザE2Eのプロセス起動処理を共通モジュールへ切り出す
relayとviteの起動・待機・停止はブラウザE2E共通の処理なので、
browser-e2e-process.mjsに移して複数のランナーから使えるようにする。
